avoid exceptions in callbacks being swallowed
Before the change, in case where an on_success callback was provided to
add_feed_to_multi or add_url_to_multi and the callback raised an error, the
error was swallowed and the on_failure callback was called without providing
any info about the exception, making it hard to find errors.

The change ensures that exceptions raised within the on_success callback
aren't silenced.